// backend/src/utils/prisma.d.ts
import { PrismaPg } from '@prisma/adapter-pg';
declare const prisma: import("@prisma/client/runtime/client").DynamicClientExtensionThis<import(".prisma/client").Prisma.TypeMap<import("@prisma/client/runtime/client").InternalArgs & {
    result: {};
    model: {};
    query: {};
    client: {};
}, {}>, import(".prisma/client").Prisma.TypeMapCb<{
    adapter: PrismaPg;
}>, {
    result: {};
    model: {};
    query: {};
    client: {};
}>;
export default prisma;
//# sourceMappingURL=prisma.d.ts.map